+++ src/Symfony/Component/Serializer/Normalizer/DataUriNormalizer.php
@@ -18,6 +18,8 @@ use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Exception\NotNormalizableValueException;
 use Symfony\Component\Serializer\InvariantViolation;
 use Symfony\Component\Serializer\Result\DenormalizationResult;
+use Symfony\Component\Serializer\Result\NormalizationResult;
+use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * Normalizes an {@see \SplFileInfo} object to a data URI.
@@ -69,10 +71,16 @@ class DataUriNormalizer implements NormalizerInterface, DenormalizerInterface, C
         }
 
         if ('text' === explode('/', $mimeType, 2)[0]) {
-            return sprintf('data:%s,%s', $mimeType, rawurlencode($data));
+            $result = sprintf('data:%s,%s', $mimeType, rawurlencode($data));
+        } else {
+            $result = sprintf('data:%s;base64,%s', $mimeType, base64_encode($data));
         }
 
-        return sprintf('data:%s;base64,%s', $mimeType, base64_encode($data));
+        if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
+            return NormalizationResult::success($result);
+        }
+
+        return $result;
     }
 
     /**
@@ -100,7 +108,7 @@ class DataUriNormalizer implements NormalizerInterface, DenormalizerInterface, C
         try {
             $result = $this->doDenormalize($data, $type);
         } catch (NotNormalizableValueException $exception) {
-            if ($context[self::COLLECT_INVARIANT_VIOLATIONS] ?? false) {
+            if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
                 $violation = new InvariantViolation($data, $exception->getMessage(), $exception);
 
                 return DenormalizationResult::failure(['' => [$violation]]);
@@ -109,7 +117,7 @@ class DataUriNormalizer implements NormalizerInterface, DenormalizerInterface, C
             throw $exception;
         }
 
-        if ($context[self::COLLECT_INVARIANT_VIOLATIONS] ?? false) {
+        if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
             return DenormalizationResult::success($result);
         }