+++ src/Symfony/Component/Serializer/Normalizer/DateIntervalNormalizer.php
@@ -15,6 +15,8 @@ use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Exception\UnexpectedValueException;
 use Symfony\Component\Serializer\InvariantViolation;
 use Symfony\Component\Serializer\Result\DenormalizationResult;
+use Symfony\Component\Serializer\Result\NormalizationResult;
+use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * Normalizes an instance of {@see \DateInterval} to an interval string.
@@ -48,7 +50,13 @@ class DateIntervalNormalizer implements NormalizerInterface, DenormalizerInterfa
             throw new InvalidArgumentException('The object must be an instance of "\DateInterval".');
         }
 
-        return $object->format($context[self::FORMAT_KEY] ?? $this->defaultContext[self::FORMAT_KEY]);
+        $result = $object->format($context[self::FORMAT_KEY] ?? $this->defaultContext[self::FORMAT_KEY]);
+
+        if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
+            return NormalizationResult::success($result);
+        }
+
+        return $result;
     }
 
     /**
@@ -80,7 +88,7 @@ class DateIntervalNormalizer implements NormalizerInterface, DenormalizerInterfa
         try {
             $result = $this->doDenormalize($data, $context);
         } catch (InvalidArgumentException | UnexpectedValueException $exception) {
-            if ($context[self::COLLECT_INVARIANT_VIOLATIONS] ?? false) {
+            if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
                 $violation = new InvariantViolation($data, $exception->getMessage(), $exception);
 
                 return DenormalizationResult::failure(['' => [$violation]]);
@@ -89,7 +97,7 @@ class DateIntervalNormalizer implements NormalizerInterface, DenormalizerInterfa
             throw $exception;
         }
 
-        if ($context[self::COLLECT_INVARIANT_VIOLATIONS] ?? false) {
+        if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
             return DenormalizationResult::success($result);
         }