+++ src/Symfony/Component/Serializer/Normalizer/DateTimeNormalizer.php
@@ -15,6 +15,8 @@ use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Exception\NotNormalizableValueException;
 use Symfony\Component\Serializer\InvariantViolation;
 use Symfony\Component\Serializer\Result\DenormalizationResult;
+use Symfony\Component\Serializer\Result\NormalizationResult;
+use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * Normalizes an object implementing the {@see \DateTimeInterface} to a date string.
@@ -47,8 +49,6 @@ class DateTimeNormalizer implements NormalizerInterface, DenormalizerInterface, 
      * {@inheritdoc}
      *
      * @throws InvalidArgumentException
-     *
-     * @return string
      */
     public function normalize($object, string $format = null, array $context = [])
     {
@@ -64,7 +64,13 @@ class DateTimeNormalizer implements NormalizerInterface, DenormalizerInterface, 
             $object = $object->setTimezone($timezone);
         }
 
-        return $object->format($dateTimeFormat);
+        $result = $object->format($dateTimeFormat);
+
+        if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
+            return NormalizationResult::success($result);
+        }
+
+        return $result;
     }
 
     /**
@@ -87,7 +93,7 @@ class DateTimeNormalizer implements NormalizerInterface, DenormalizerInterface, 
         try {
             $result = $this->doDenormalize($data, $type, $context);
         } catch (NotNormalizableValueException $exception) {
-            if ($context[self::COLLECT_INVARIANT_VIOLATIONS] ?? false) {
+            if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
                 $violation = new InvariantViolation($data, $exception->getMessage(), $exception);
 
                 return DenormalizationResult::failure(['' => [$violation]]);
@@ -96,7 +102,7 @@ class DateTimeNormalizer implements NormalizerInterface, DenormalizerInterface, 
             throw $exception;
         }
 
-        if ($context[self::COLLECT_INVARIANT_VIOLATIONS] ?? false) {
+        if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
             return DenormalizationResult::success($result);
         }